Implemented Adjusted Designs
There have been some layout changes to the source, as well as the inclusion of a new asset which were decided upon between commits. This includes the addition of an image in the first couple of frames, as well as having the details and preview pages both be on the same frame. The details view frame is set up, but code does still need to be made to make the table of data.